requirements.txt
streamlit
pytesseract
pokemontcgsdk
opencv-python-headless
pillow